pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "geovisualizer"
version = "0.0.1"
authors = [
  { name="Valentin Bruch", email="software@vbruch.eu" },
]
description = "Basic 3d visualization of GPX files"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
  "Programming Language :: Python :: 3",
  "Development Status :: 2 - Pre-Alpha",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Games/Entertainment",
  "Natural Language :: English",
]
dependencies = [
  "numpy",
  "pyqtgraph",
  "lxml",
  "pyproj",
  "Pillow",
]
